requirements.txt
langchain-core
langchain-community
langchain-ollama
pypdf
wikipedia
langchain-chroma>=0.1.2